Add new Customer Login window
Window is currently split with an existing customer login on the left
and new customer login on the right.

Alternatively I could make the page default to existing customer and
load a new page for a new customer.

Couldn't think of a better way to implement adding a user's insurance
reference number and drivers license class than just in a form, so
that's what it uses for now.

Next steps: Finish touching up the details in the text boxes and
buttons, and implement writing new customers to database with the New
Customer section.