:pencil2: Suggestions for layout edits
There were a couple of things I  changed here, one was a section header
that was probably a copy-paste, I also edited the form title across the
top bar.

NOTE: I left the Renter Age text box but we already have that question
in the customer account creation page. Not sure how you guys want to
handle it but I don't think we should have it in both places. Let me
know.