:lock: Add functions for checking passwords
When logging in, we can write a query to obtain a hashed password stored
in the database associated with a specific email account. The
user-entered password can be hashed and compared with the stored hashed
password. If they are equal, we can authenticate the user.
No salting is used for simplicity (unless Microsoft's built in methods
do this) because it is not going to be deployed anywhere.

+ new private method Login.HashPassword(string)
+ new private method Login.VerifyPassword(string)