:sparkles: Login Window updates Main form
Now the Login window acts as a temporary pop-up to log in a new user and
save/retrieve their details for the main page.

This means that the customer-specific details can be queried for in the
subsequent pages as long as it is passed to the next window as a
parameter.

WIP: Does not yet work for existing customer login; login must query
database for user information and obtain user ID.